feat(signals): add Vercel crons for queue refresh and orchestration
Wire scheduler-driven endpoints so discussion refresh can enqueue signal
evaluations and the orchestrator can drain due queue items automatically.
