+++ apps/web/.env.template
@@ -102,6 +102,16 @@ MATRIX_REGISTRATION_SHARED_SECRET=
 MATRIX_PASSWORD_SECRET=
 # Optional bot token for org memory / MCP Matrix reads (server-only).
 # HYPHA_MATRIX_ORG_MEMORY_ACCESS_TOKEN=
+# Ops secret for Space Memory / signal orchestrator admin routes (x-hypha-ops-secret or Bearer).
+# HYPHA_SPACE_MEMORY_OPS_SECRET=
+# Vercel Cron: set in the Vercel project; Vercel sends Authorization: Bearer on cron hits.
+# CRON_SECRET=
+# Service token used when the signal orchestrator reads org memory (falls back to HYPHA_MCP_AUTH_TOKEN).
+# HYPHA_SIGNAL_ORCHESTRATOR_AUTH_TOKEN=
+# HYPHA_MCP_AUTH_TOKEN=
+# HYPHA_MCP_MATRIX_REQUEST_URL=
+# Debounce window before a queued signal evaluation becomes due (minutes).
+# HYPHA_SIGNAL_ORCHESTRATOR_WINDOW_MINUTES=20
 # When TURN/coturn uses a different host than the homeserver, add CSP connect-src origins (comma-separated).
 # Example: https://srv1294735.hstgr.cloud — see docs/requirements/matrix-voip-turn-server-setup.md
 # NEXT_PUBLIC_MATRIX_TURN_CONNECT_SOURCES=
